+++ apps/kobayashi/include/kobayashi/verification_utilities.hpp
@@ -17,7 +17,7 @@
 #include <openpfc/frontend/io/png_writer.hpp>
 #include <openpfc/kernel/data/grid_field.hpp>
 #include <openpfc/kernel/decomposition/decomposition.hpp>
-#include <openpfc_apps/gather.hpp>
+#include <kobayashi/xy_gather.hpp>
 
 namespace {